requirements.txt
beautifulsoup4==4.14.2
fasttext_wheel==0.9.2
huggingface_hub==0.35.1
python-dotenv==1.1.1
pytz==2025.2
Requests==2.32.5
sumy==0.11.0
telethon==1.41.2
